Cleaning up dirs + adding rest of CRUD operations
